// src/app/page/master/map-kelompoktransaksi-ruangan/map-kelompoktransaksi-ruangan.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapKelompoktransaksiRuangan {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	value?: any;
	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}